add option to push changes to users on creation

+++ manifests/bash/custom_ps1.pp
@@ -51,13 +51,15 @@ class cluster_tools::bash::custom_ps1(
   }
 
   # Copy to the existing users only when the file is overwritten by puppet
-  exec { 'push updated .bashrc and .profile to the existing users':
-    command     => 'sudo find /home -type d -exec cp /etc/skel/.bashrc {} \; -exec cp /etc/skel/.profile {} \;',
-    path        => $cluster_tools::path,
-    subscribe   => [
-      File['/etc/skel/.bashrc'],
-      File['/etc/skel/.profile'],
-    ],
-    refreshonly => true,
+  if $user_onetime_overwrite {
+    exec { 'push updated .bashrc and .profile to the existing users':
+      command     => 'sudo find /home -type d -exec cp /etc/skel/.bashrc {} \; -exec cp /etc/skel/.profile {} \;',
+      path        => $cluster_tools::path,
+      subscribe   => [
+        File['/etc/skel/.bashrc'],
+        File['/etc/skel/.profile'],
+      ],
+      refreshonly => true,
+    }
   }
 }
